refactor : step2 진행을 위한 패키지 구조 변경
