+++ src/main/java/cart/controller/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package cart.product.dto;
+package cart.controller.dto;
 
 public class ExceptionResponse {
